refactor: 웹 클라이언트 자체를 글로벌에 인터페이스로 변경
