+++ chatbot/backend/src/main/java/com/hallachatbot/backend/global/client/service/AiServiceClientImpl.java
@@ -1,4 +1,4 @@
-package com.hallachatbot.backend.global.client;
+package com.hallachatbot.backend.global.client.service;
 
 import java.time.Duration;
 import java.util.List;
@@ -19,33 +19,16 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
 
-/**
- * AI 서비스 연동 클라이언트
- *
- * <p>
- * Python의 httpx 로직을 WebClient로 이식<br>
- * SSE(또는 NDJSON) 스트림을 받아 처리함
- * </p>
- *
- * @author pwk0131
- */
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class AiServiceClient {
+public class AiServiceClientImpl implements AiServiceClient {
 
 	private final WebClient.Builder webClientBuilder;
 
 	@Value("${app.ai-service-url}")
 	private String aiServiceUrl;
 
-	/**
-	 * AI 챗봇 스트리밍 요청
-	 *
-	 * @param request 사용자 요청 정보
-	 * @param history 대화 히스토리
-	 * @return AI 응답 스트림 (Flux)
-	 */
 	public Flux<AiServiceResponse> streamChat(ChatRequest request, List<ChatHistoryResponse> history) {
 		String endpoint = aiServiceUrl + "/api/chat";
